requirements.txt
streamlit>=1.35
requests>=2.31
pandas>=2.2
numpy>=1.26
plotly>=5.22
python-dotenv>=1.0
tqdm>=4.66
